laravel assign till day11

+++ app/Http/Controllers/Categories/CategoryController.php
@@ -3,31 +3,42 @@
 namespace App\Http\Controllers\Categories;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class CategoryController extends Controller {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
     public function index() {
+        if(!Gate::allows('category_list')){
+            return abort(401);
+        }
         $categories = Category::all();
         // dd($categories);
         return view('categories.index', compact('categories'));
     }
     public function create() {
+        if(!Gate::allows('category_create')){
+            return abort(401);
+        }
         return view('categories.create');
     }
-    public function store(Request $request) {
+    public function store(CategoryRequest $request) {
         // dd($request->all());
-        $request->validate([
-            'name' => ['required', 'string'],
-            'description' => ['required','string'],
-            'image' => ['required', 'image'],
-         'status' => ['boolean'],
-        ]);
+        if(!Gate::allows('category_create')){
+            return abort(401);
+        }
+        $data = $request->validated();
+        // dd($data);
         // $path = $request->file('image')->store('category');
         // dd($request->image);
-        $imageName = time().'.'.$request->image->getClientOriginalExtension();
-
+        $imageName = time().'.'.$data['image']->getClientOriginalExtension();
+        // dd($imageName);
         $request->image->move(public_path('/uploadedimages'), $imageName);
         Category::create([
             'name' => $request->name,
@@ -45,12 +56,20 @@ class CategoryController extends Controller {
     }
     public function edit($id) {
         // dd($id);
+        if(!Gate::allows('category_edit')){
+            return abort(401);
+        }
         $category = Category::find($id);
         $category = Category::where('id', $id)->first();
         // dd($category);
         return view('categories.edit', compact('category'));
     }
     public function update(Request $request, $id) {
+
+        if(!Gate::allows('category_edit')){
+            return abort(401);
+        }
+
         $category = Category::find($id);
         // dd($category);
         $category->update([
@@ -67,6 +86,11 @@ class CategoryController extends Controller {
         return redirect()->route('categories.index');
     }
     public function delete($id) {
+
+        if(!Gate::allows('category_delete')){
+            return abort(401);
+        }
+
         Category::where('id', $id)->delete();
         // DB::table('categories')->where('id', $id)->delete();
         return redirect()->route('categories.index');
